feat: rename package from aiuda-planner-agent to datapilot
- Rename src/aiuda_planner/ to src/datapilot/
- Update package name in pyproject.toml
- Update CLI entry point to 'datapilot' command
- Update all internal imports from aiuda_planner to datapilot
- Update environment variable prefix from AIUDA_* to DATAPILOT_*
- Update README with new package name and examples
- Update all examples to use new package name
- Update all tests to use new imports

This rename makes the package more community-friendly for
distribution as an open-source Python package.

Closes #8
