+++ .env.example
@@ -1,5 +1,5 @@
 # =============================================================================
-# Aiuda Planner Agent - Configuration
+# DataPilot - Configuration
 # =============================================================================
 # Copy this file to .env and fill in your values
 # Priority: CLI args > Environment variables > .env file > defaults
@@ -54,16 +54,16 @@ GOOGLE_API_KEY=your-google-api-key-here
 # =============================================================================
 
 # Maximum rounds before stopping (default: 30)
-# AIUDA_MAX_ROUNDS=30
+# DATAPILOT_MAX_ROUNDS=30
 
 # LLM temperature (default: 0.3)
-# AIUDA_TEMPERATURE=0.3
+# DATAPILOT_TEMPERATURE=0.3
 
 # Max tokens per response (default: 4096)
-# AIUDA_MAX_TOKENS=4096
+# DATAPILOT_MAX_TOKENS=4096
 
 # Code execution timeout in seconds (default: 300)
-# AIUDA_CODE_TIMEOUT=300
+# DATAPILOT_CODE_TIMEOUT=300
 
 # Default workspace directory (default: ./workspace)
-# AIUDA_WORKSPACE=./workspace
+# DATAPILOT_WORKSPACE=./workspace
