--- src/aiuda_planner/core/__init__.py
@@ -1,11 +1,0 @@
-"""Core modules for the AI Planner Agent."""
-
-from aiuda_planner.core.executor import JupyterExecutor
-from aiuda_planner.core.planner import PlanParser
-from aiuda_planner.core.engine import AgentEngine
-
-__all__ = [
-    "JupyterExecutor",
-    "PlanParser",
-    "AgentEngine",
-]
