+++ src/datapilot/__init__.py
@@ -8,14 +8,14 @@ This package provides a powerful agent for data analysis tasks that:
 - Streams events for real-time UI updates
 
 Basic usage:
-    from aiuda_planner import PlannerAgent
+    from datapilot import PlannerAgent
 
     with PlannerAgent(model="gpt-4o") as agent:
         result = agent.run("Analyze sales data and find trends")
         print(result.answer)
 
 With streaming:
-    from aiuda_planner import PlannerAgent, EventType
+    from datapilot import PlannerAgent, EventType
 
     agent = PlannerAgent()
     for event in agent.run_stream("Build a model"):
@@ -23,8 +23,8 @@ With streaming:
             print(f"Code: {event.success}")
 """
 
-from aiuda_planner.agents.base import PlannerAgent, AgentResult
-from aiuda_planner.schema.models import (
+from datapilot.agents.base import PlannerAgent, AgentResult
+from datapilot.schema.models import (
     AgentConfig,
     AgentEvent,
     EventType,
@@ -33,13 +33,13 @@ from aiuda_planner.schema.models import (
     PlanStep,
     SessionState,
 )
-from aiuda_planner.core.executor import JupyterExecutor
-from aiuda_planner.core.engine import AgentEngine
-from aiuda_planner.core.planner import PlanParser
-from aiuda_planner.core.context import RunContext
-from aiuda_planner.utils.logger import AgentLogger, Colors
-from aiuda_planner.utils.notebook import NotebookBuilder, ExecutionTracker
-from aiuda_planner.utils.run_logger import RunLogger
+from datapilot.core.executor import JupyterExecutor
+from datapilot.core.engine import AgentEngine
+from datapilot.core.planner import PlanParser
+from datapilot.core.context import RunContext
+from datapilot.utils.logger import AgentLogger, Colors
+from datapilot.utils.notebook import NotebookBuilder, ExecutionTracker
+from datapilot.utils.run_logger import RunLogger
 
 __version__ = "0.3.0"
