+++ src/datapilot/agents/__init__.py
@@ -1,6 +1,6 @@
 """Agent implementations for the AI Planner."""
 
-from aiuda_planner.agents.base import PlannerAgent
+from datapilot.agents.base import PlannerAgent
 
 __all__ = [
     "PlannerAgent",
