+++ src/datapilot/agents/base.py
@@ -7,21 +7,21 @@ from pathlib import Path
 from typing import Optional, Callable, Any, Generator, TYPE_CHECKING
 from datetime import datetime
 
-from aiuda_planner.schema.models import (
+from datapilot.schema.models import (
     AgentConfig,
     AgentEvent,
     EventType,
     SessionState,
     PlanState,
 )
-from aiuda_planner.core.executor import JupyterExecutor
-from aiuda_planner.core.engine import AgentEngine
-from aiuda_planner.utils.logger import AgentLogger
-from aiuda_planner.utils.notebook import NotebookBuilder
+from datapilot.core.executor import JupyterExecutor
+from datapilot.core.engine import AgentEngine
+from datapilot.utils.logger import AgentLogger
+from datapilot.utils.notebook import NotebookBuilder
 
 if TYPE_CHECKING:
-    from aiuda_planner.core.context import RunContext
-    from aiuda_planner.utils.run_logger import RunLogger
+    from datapilot.core.context import RunContext
+    from datapilot.utils.run_logger import RunLogger
 
 
 class PlannerAgent:
@@ -110,7 +110,7 @@ class PlannerAgent:
         # Initialize run logger if context provided
         self._run_logger: Optional["RunLogger"] = None
         if context:
-            from aiuda_planner.utils.run_logger import RunLogger
+            from datapilot.utils.run_logger import RunLogger
             self._run_logger = RunLogger(context)
 
         # Initialize components
