+++ src/datapilot/core/__init__.py
@@ -1,0 +1,11 @@
+"""Core modules for the AI Planner Agent."""
+
+from datapilot.core.executor import JupyterExecutor
+from datapilot.core.planner import PlanParser
+from datapilot.core.engine import AgentEngine
+
+__all__ = [
+    "JupyterExecutor",
+    "PlanParser",
+    "AgentEngine",
+]
