+++ src/datapilot/utils/__init__.py
@@ -1,7 +1,7 @@
 """Utility modules for the AI Planner Agent."""
 
-from aiuda_planner.utils.logger import AgentLogger
-from aiuda_planner.utils.notebook import NotebookBuilder, ExecutionTracker
+from datapilot.utils.logger import AgentLogger
+from datapilot.utils.notebook import NotebookBuilder, ExecutionTracker
 
 __all__ = [
     "AgentLogger",
